chore: clean and add chackraui
