+++ package.json
@@ -8,6 +8,10 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@chakra-ui/react": "^1.7.3",
+    "@emotion/react": "^11",
+    "@emotion/styled": "^11",
+    "framer-motion": "^4",
     "next": "12.0.7",
     "react": "17.0.2",
     "react-dom": "17.0.2"
