+++ src/pages/_app.tsx
@@ -1,0 +1,14 @@
+// import '../styles/globals.css'
+import type { AppProps } from 'next/app'
+
+import { ChakraProvider } from '@chakra-ui/react'
+
+function MyApp({ Component, pageProps }: AppProps) {
+  return (
+    <ChakraProvider>
+      <Component {...pageProps} />
+    </ChakraProvider>
+  )
+}
+
+export default MyApp
