+++ src/pages/index.tsx
@@ -1,0 +1,17 @@
+import type { NextPage } from 'next'
+import Head from 'next/head'
+
+const Home: NextPage = () => {
+  return (
+    <div>
+      <Head>
+        <title>Create Next App</title>
+        <meta name="description" content="Generated by create next app" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+      <h1>Hello</h1>
+    </div>
+  )
+}
+
+export default Home
